requirements.txt
pymongo==4.16.0
dnspython==2.8.0
python-dotenv==1.2.1
requests==2.32.5
google-cloud-storage==3.9.0
google-auth==2.49.1
pyspark==3.5.5
pandas==3.0.1
numpy==2.4.1
apache-airflow==3.1.6
